+++ test/libsolidity/syntaxTests/nameAndTypeResolution/039_functions_with_identical_structs_in_interface.sol
@@ -1,13 +1,11 @@
 pragma experimental ABIEncoderV2;
 
 contract C {
-    struct S1 { }
-    struct S2 { }
+    struct S1 { int i; }
+    struct S2 { int i; }
     function f(S1) pure {}
     function f(S2) pure {}
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
-// Warning: (52-65): Defining empty structs is deprecated.
-// Warning: (70-83): Defining empty structs is deprecated.
-// TypeError: (115-137): Function overload clash during conversion to external types for arguments.
+// TypeError: (129-151): Function overload clash during conversion to external types for arguments.